+++ Resulti.Core/Util/NotificationType.cs
@@ -1,0 +1,25 @@
+namespace Resulti.Core.Util
+{
+    /// <summary>
+    /// Utility for notification type names
+    /// </summary>
+    public static class NotificationType
+    {
+        /// <summary>
+        /// Designation for a information notification
+        /// </summary>
+        public const string Info = "INFO";
+        /// <summary>
+        /// Designation for a success notification
+        /// </summary>
+        public const string Success = "SUCCESS";
+        /// <summary>
+        /// Designation for a warning notification
+        /// </summary>
+        public const string Warning = "WARNING";
+        /// <summary>
+        /// Designation for an error notification
+        /// </summary>
+        public const string Error = "ERROR";
+    }
+}