+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ### V0.1.0
 
 + Initial version
+
+### V0.1.1
+
++ Added result builder functionality
++ Added result extension methods
++ Added type to notifications
++ Added notification type utility